+++ KeyboardKitDemo/AppDelegate.swift
@@ -16,7 +16,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate, KeyboardStateInspector {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         let isEnabled = isKeyboardEnabled(for: "com.danielsaidi.keyboardkit.demo.keyboard")
-        print(isEnabled ? "Keyboard is enabled" : "Keyboard is disabled")
+        let isEnabledSwiftUI = isKeyboardEnabled(for: "com.danielsaidi.keyboardkit.demo.swiftui")
+        print(isEnabled ? "UIKit Demo Keyboard is enabled" : "UIKit Demo Keyboard is disabled")
+        print(isEnabledSwiftUI ? "SwiftUI Demo Keyboard is enabled" : "SwiftUI Demo Keyboard is disabled")
         return true
     }